+++ src/main/java/org/blue/webframework/sys/account/service/impl/PrivilegeServiceImpl.java
@@ -23,9 +23,12 @@ class PrivilegeServiceImpl implements PrivilegeService {
 	private RoleResourceMapper roleResourceMapper;
 	
 	@Override
-	public void grantAccount(long accountId, String operate, long resourceId) {
+	public void grantAccount(int accountId, String operate, int resourceId) {
+		//判断该账号是否有权限
 		if(accountResourceMapper.getCount(accountId, operate, resourceId)<=0)
 		{
+			//否
+			//给账号授权
 			AccountResource accountResource=new AccountResource();
 			accountResource.setOperate(operate);
 			accountResource.setResourceId(resourceId);
@@ -35,15 +38,17 @@ class PrivilegeServiceImpl implements PrivilegeService {
 	}
 
 	@Override
-	public void grantAccount(long accountId, String operate, String resourceCode) {
+	public void grantAccount(int accountId, String operate, String resourceCode) {
 		Resource resource=resourceMapper.selectByCode(resourceCode);
 		grantAccount(accountId,operate,resource.getId());
 	}
 
 	@Override
-	public void grantRole(long roleId, String operate, long resourceId) {
+	public void grantRole(int roleId, String operate, int resourceId) {
+		//判断角色信息是否存在
 		if(roleResourceMapper.getCount(roleId, operate, resourceId)<=0)
 		{
+			//不存在，执行添加
 			RoleResource roleResource=new RoleResource();
 			roleResource.setOperate(operate);
 			roleResource.setResourceId(resourceId);
@@ -51,12 +56,14 @@ class PrivilegeServiceImpl implements PrivilegeService {
 			roleResourceMapper.insert(roleResource);
 		}
 			
-		
 	}
 
 	@Override
-	public void grantRole(long roleId, String operate, String resourceCode) {
+	public void grantRole(int roleId, String operate, String resourceCode) {
+		//根据资源编码查询资源信息
 		Resource resource=resourceMapper.selectByCode(resourceCode);
+		//资源信息已存在，则不添加。
+		//不存在，则执行添加资源信息
 		grantRole(roleId,operate,resource.getId());
 		
 	}
@@ -64,23 +71,23 @@ class PrivilegeServiceImpl implements PrivilegeService {
 	
 	
 	@Override
-	public boolean hasAccountPermission(long accountId, String operate, String resourceCode) {
+	public boolean hasAccountPermission(int accountId, String operate, String resourceCode) {
 		return roleResourceMapper.getCountWidthCode(accountId, operate, resourceCode)>0;
 	}
 
 	@Override
-	public boolean hasAccountPermissionWithName(long accountId, String operate, String resourceName) {
+	public boolean hasAccountPermissionWithName(int accountId, String operate, String resourceName) {
 		return accountResourceMapper.getCountWidthName(accountId, operate, resourceName)>0;
 	}
 
 	@Override
-	public boolean hasRolePermission(long roleId, String operate, String resourceCode) {
+	public boolean hasRolePermission(int roleId, String operate, String resourceCode) {
 
 		return accountResourceMapper.getCountWidthCode(roleId, operate, resourceCode)>0;
 	}
 
 	@Override
-	public boolean hasRolePermissionWithName(long roleId, String operate, String resourceName) {
+	public boolean hasRolePermissionWithName(int roleId, String operate, String resourceName) {
 		return roleResourceMapper.getCountWidthName(roleId, operate, resourceName)>0;
 	}
 	
@@ -94,13 +101,13 @@ class PrivilegeServiceImpl implements PrivilegeService {
 	}
 
 	@Override
-	public void revokeAccount(long accountId, String operate, long resourceId) {
+	public void revokeAccount(int accountId, String operate, int resourceId) {
 		accountResourceMapper.delete(accountId, operate, resourceId);
 
 	}
 
 	@Override
-	public void revokeRole(long roleId, String operate, long resourceId) {
+	public void revokeRole(int roleId, String operate, int resourceId) {
 		roleResourceMapper.delete(roleId, operate, resourceId);
 	}
 }
